rename utils folder
